removed all dependencies from the javascript test (contract test no longer coupled with it)
