+++ client/client-contract-test/test/client-sdk.spec.ts
@@ -14,6 +14,7 @@ import * as path from "path";
 import { OrbsClient, OrbsContract, Address } from "../../client-sdk-javascript/src";
 import { OrbsContractMethodArgs } from "../../client-sdk-javascript/src/orbs-contract";
 import { expectedCallContractRequest, expectedSendTransactionRequest, SENDER_ADDRESS, CONTRACT_NAME, CONTRACT_METHOD_NAME, CONTRACT_METHOD_ARGS, SENDER_PUBLIC_KEY, VIRTUAL_CHAIN_ID } from "../src/expected-results"
+import { testContract } from "../src/contract-adapter";
 
 chai.use(sinonChai);
 
@@ -161,29 +162,3 @@ after((done) => {
   // httpServer.close(done);
   done();
 });
-
-function testContract(makeContract: () => OrbsContractAdapter) {
-
-    describe("calls the connector interface with the correct inputs when", async function () {
-
-      it("getSendTranscationObject() is called", async () => {
-        const sendTransactionObject = await makeContract().getSendTranscationObject();
-        expect(sendTransactionObject).to.have.property("payload").that.is.eql(expectedSendTransactionRequest.payload);
-        expect(sendTransactionObject).to.have.property("header").that.has.property("senderAddressBase58").that.is.eql(expectedSendTransactionRequest.header.senderAddressBase58);
-        expect(sendTransactionObject).to.have.property("header").that.has.property("contractAddressBase58").that.is.eql(expectedSendTransactionRequest.header.contractAddressBase58);
-        expect(sendTransactionObject).to.have.property("header").that.has.property("timestamp").that.is.a("string");
-        const now = Date.now();
-        // testing that timestamp is less than a couple of seconds old and not in the future
-        const timestamp = Number(sendTransactionObject.header.timestamp);
-        expect(timestamp).to.be.above(now - (30 * 1000));
-        expect(timestamp).to.be.below(now + (10 * 1000));
-      });
-
-      it("getCallObject() is called", async () => {
-        const callObject = await makeContract().getCallObject();
-        expect(callObject).to.have.property("payload").that.is.eql(expectedCallContractRequest.payload);
-        expect(callObject).to.have.property("contractAddressBase58", expectedCallContractRequest.contractAddressBase58);
-        expect(callObject).to.have.property("senderAddressBase58", expectedCallContractRequest.senderAddressBase58);
-      });
-    });
-}
