+++ client/client-sdk-javascript/package.json
@@ -16,24 +16,16 @@
   },
   "devDependencies": {
     "@types/bindings": "^1.3.0",
-    "@types/bluebird": "^3.5.20",
     "@types/chai": "^4.1.2",
     "@types/chai-as-promised": "^7.1.0",
-    "@types/express": "^4.11.1",
-    "@types/java": "^0.7.32",
     "@types/mocha": "^2.2.47",
-    "@types/nock": "^9.1.3",
-    "@types/node": "^9.4.4",
     "@types/request-promise": "^4.1.41",
-    "@types/sinon": "^4.3.1",
+    "@types/sinon": "^4.3.2",
     "@types/sinon-chai": "^2.7.29",
-    "@types/supertest": "^2.0.4",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "concurrently": "^3.5.1",
     "elliptic": "^6.4.0",
-    "express": "^4.16.3",
-    "java": "^0.9.1",
     "mocha": "^5.0.1",
     "node-gyp": "^3.6.2",
     "request-promise": "^4.2.2",
@@ -41,14 +33,10 @@
     "sinon-chai": "^3.0.0",
     "ts-sinon": "https://github.com/orbs-network/ts-sinon",
     "tslint": "^5.9.1",
-    "typescript": "^2.7.1"
+    "typescript": "^2.8.1"
   },
   "dependencies": {
-    "@types/morgan": "^1.7.35",
     "bindings": "^1.3.0",
-    "bluebird": "^3.5.1",
-    "orbs-interfaces": "file:../../projects/architecture",
-    "path": "^0.12.7",
-    "python-bridge": "^1.0.3"
+    "orbs-interfaces": "file:../../projects/architecture"
   }
 }
